requirements.txt
Django==1.11.7
djangorestframework==3.7.3
psycopg2==2.6.1
python-decouple==3.0
gunicorn==19.7.1
Pillow==3.1.1
ipdb==0.9.0
requests==2.13.0
social-auth-app-django==1.1.0
python-memcached==1.58
